refactor(github): separate pull request file errors
