+++ src/features/github/github-pull-request-file-view-state.tsx
@@ -1,6 +1,3 @@
-import { CircleAlert, RefreshCw } from "lucide-react";
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import { Button } from "@/components/ui/button";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Spinner } from "@/components/ui/spinner";
 import { useAppTranslation } from "@/hooks/use-app-translation";
@@ -24,32 +21,6 @@ export function hasUnmatchedPullRequestFileViewStates(
   if (!snapshot) return false;
   const statePaths = new Set(snapshot.files.map((file) => file.path));
   return files.some((file) => !statePaths.has(file.path));
-}
-
-export function GitHubPullRequestFilesErrorAlert({
-  title,
-  message,
-  actionLabel,
-  onAction,
-}: {
-  title: string;
-  message: string;
-  actionLabel: string;
-  onAction: () => void;
-}) {
-  return (
-    <Alert variant="destructive">
-      <CircleAlert />
-      <AlertTitle>{title}</AlertTitle>
-      <AlertDescription className="flex flex-wrap items-center justify-between gap-2">
-        <span>{message}</span>
-        <Button type="button" variant="outline" size="xs" onClick={onAction}>
-          <RefreshCw data-icon="inline-start" />
-          {actionLabel}
-        </Button>
-      </AlertDescription>
-    </Alert>
-  );
 }
 
 export function GitHubPullRequestFileViewCheckbox({
